Update books using PATCH (#12)
Summary:
This section looks at updating books using PATCH.
We tested the books can be updated one property at a time or all at
once. If a book does not exist or a property is in the wrong format, it
should not update anything and return the appropriate status code.

In terms of the file structure, I have put this function in its own file
and the corresponding tests in separate file too. I also extracted the
`books` variable into its own file as it is referenced in this `update`
function, as well as the `app.js` file. I assume other functions would
use it too, so it would be good to reference it rather than duplicate it
in each file.

Co-author: Korinna Pinakoulaki

To do:
Remove the existing `books` array in the `app.js` file to use the 'new'
one.